requirements.txt
fastapi
uvicorn
xgboost
joblib
pytesseract
Pillow
malaya
numpy
beautifulsoup4
torch --extra-index-url https://download.pytorch.org/whl/cp